chore: move insertionindex into store pkg
